Split end-to-end flow instantiation into focused collaborators
CreateEndToEndFlowsSwitch combined the instance traversal, per-component
candidate discovery, nested-flow expansion, diagnostics, commit,
connection matching, and mode computation in about 1600 lines. Separate
those responsibilities so the traversal and atomicity invariants can be
reviewed on their own.

The switch stays the public entry point and keeps every protected
extension point with its current signature: it visits component
instances, resolves their implementation through the classifier cache,
watches the progress monitor, reports errors, counts clone names, and
holds the active session. The new internal collaborators are:

- EndToEndFlowSession owns one component instance's discovery, which is
  the former FlowInstantiationContext merged with the traversal, the
  candidate and traversal-state types, the deferred diagnostics, and the
  atomic commit.
- FlowConnectionMatcher holds the stateless connection lookup and the
  compatibility predicates.
- EndToEndFlowModes holds mode resolution and system-operation-mode
  computation.
- FlowIterator and ETEInfo move out of the switch unchanged in behavior.

Extension behavior is preserved rather than reimplemented. The session
never calls its own traversal steps or the mode computation directly;
both go out through FlowInstantiationHost, the switch's protected
methods seen from inside, implemented by a private inner class that
forwards to them. An override therefore still decides every step, and
resolving the session for a protected step fails with the same
IllegalStateException as before when there is no instantiation in
progress.

The moved method bodies are unchanged apart from reading session fields
directly, dropping the context parameter, taking the nested candidate's
connections and instance instead of the candidate, and collecting flow
implementations in an ArrayList instead of a BasicEList. ETEInfo is now
built by the session with final fields; it remains unreachable outside
the bundle, as it was when it was a package-private member of the
switch, and the same is true of FlowIterator and of the protected
methods whose signatures name either type.

doc/e2e_instantiation.md gains the ownership boundaries and the type
table; the internal package documentation gains the flow discovery
paragraph.
